Combine args helpers
